alg.cpp

+++ src/alg.cpp
@@ -4,7 +4,7 @@
 
 double pown(double value, uint16_t n) {
   double result = value;
-  for (int i = 0; i < n; i++) {
+  for (uint16_t i = 0; i < n; i++) {
     result = result * result;
   }
   return result;
@@ -12,7 +12,7 @@ double pown(double value, uint16_t n) {
 
 uint64_t fact(uint16_t n) {
   uint64_t result = 1, a = 1;
-  for (int i = 0; i < n; i++) {
+  for (uint16_t i = 0; i < n; i++) {
     result = result * a;
     a++;
   }
@@ -23,7 +23,7 @@ double calcItem(double x, uint16_t n) { return (pown(x, n) / fact(n)); }
 
 double expn(double x, uint16_t count) {
   double result;
-  for (int i = 0; i < count; i++) {
+  for (uint16_t i = 0; i < count; i++) {
     result += calcItem(x, i);
   }
   return result;
@@ -31,7 +31,7 @@ double expn(double x, uint16_t count) {
 
 double sinn(double x, uint16_t count) {
   double result = 0;
-  for (int i = 0; i < count; i++) {
+  for (uint16_t i = 0; i < count; i++) {
     result += pown(-1, i) * pown(x, 2 * i + 1) / fact(2 * i);
   }
   return result;
@@ -39,7 +39,7 @@ double sinn(double x, uint16_t count) {
 
 double cosn(double x, uint16_t count) {
   double result = 0;
-  for (int i = 0; i < count; i++) {
+  for (uint16_t i = 0; i < count; i++) {
     result += pown(-1, i - 1) * pown(x, 2 * i - 2) / fact(2 * i - 2);
   }
   return result;
